remove action type

+++ Music-Bot-Telegram.Data/Models/User.cs
@@ -14,7 +14,7 @@ public class User : IEntity
     {
     }
     
-    public User(int id)
+    public User(long id)
     {
         Id = id;
     }
